- default CTRLP mapping
`c` meant that it always searched in the current directory which isn't
ideal.

`ra` is the default and `ra` means that it will traverse up directories
till it finds a .git/.svn (and others) file and then search from there.

This still presents issues when your project has submodules and you're
working in a submodule but it's better than it was.